feat: change JSON request body handling
Handle JSON request body in an opinionated fasion that allows trustful
detection in a long term way and let undici handles all other
types.

FTR: old implementation was handling FormData as JSON and prevent
undici to handle it properly.

BREAKING CHANGE: Header  will not be automagically added anymore, you will need to wrap it with a call to  function.